clean up and organize
