+++ packages/manager/src/features/ObjectStorage/BucketDetail/CreateFolderDrawer.tsx
@@ -4,7 +4,7 @@ import React, { useEffect } from 'react';
 import { ActionsPanel } from 'src/components/ActionsPanel/ActionsPanel';
 import { Drawer } from 'src/components/Drawer';
 import { TextField } from 'src/components/TextField';
-import { useCreateObjectUrlMutation } from 'src/queries/objectStorage';
+import { useCreateObjectUrlMutation } from 'src/queries/object-storage/queries';
 
 interface Props {
   bucketName: string;
