+++ packages/manager/src/features/ObjectStorage/BucketLanding/BucketLanding.tsx
@@ -15,7 +15,7 @@ import { useOpenClose } from 'src/hooks/useOpenClose';
 import {
   useDeleteBucketMutation,
   useObjectStorageBuckets,
-} from 'src/queries/objectStorage';
+} from 'src/queries/object-storage/queries';
 import { useProfile } from 'src/queries/profile/profile';
 import { useRegionsQuery } from 'src/queries/regions/regions';
 import {
@@ -35,7 +35,7 @@ import type {
   ObjectStorageCluster,
 } from '@linode/api-v4';
 import type { Theme } from '@mui/material/styles';
-import type { BucketError } from 'src/queries/objectStorage';
+import type { BucketError } from 'src/queries/object-storage/requests';
 
 const useStyles = makeStyles()((theme: Theme) => ({
   copy: {
