+++ packages/manager/src/features/ObjectStorage/BucketLanding/ClusterSelect.tsx
@@ -2,7 +2,7 @@ import { Region } from '@linode/api-v4/lib/regions';
 import * as React from 'react';
 
 import { RegionSelect } from 'src/components/RegionSelect/RegionSelect';
-import { useObjectStorageClusters } from 'src/queries/objectStorage';
+import { useObjectStorageClusters } from 'src/queries/object-storage/queries';
 import { useRegionsQuery } from 'src/queries/regions/regions';
 
 interface Props {
