+++ packages/manager/src/features/ObjectStorage/BucketLanding/OveragePricing.tsx
@@ -6,7 +6,7 @@ import { CircleProgress } from 'src/components/CircleProgress';
 import { TextTooltip } from 'src/components/TextTooltip';
 import { Typography } from 'src/components/Typography';
 import { useNetworkTransferPricesQuery } from 'src/queries/networkTransfer';
-import { useObjectStorageTypesQuery } from 'src/queries/objectStorage';
+import { useObjectStorageTypesQuery } from 'src/queries/object-storage/queries';
 import { UNKNOWN_PRICE } from 'src/utilities/pricing/constants';
 import { getDCSpecificPriceByType } from 'src/utilities/pricing/dynamicPricing';
